unsigned - signed comparison compiler remarks

+++ src/doc.h
@@ -23,7 +23,7 @@ public:
   const vector<int>& get_ws() const {return ws;}
   
   const int get_w(int i) const {
-	assert(i < ws.size());
+	assert(i < ((int)ws.size()));
 	return ws[i];
   }
